+++ gallery/identity/ceed-identity.c
@@ -30,6 +30,8 @@ static int CeedQFunctionInit_Identity(Ceed ceed, const char *requested,
   // QFunction fields 'input' and 'output' with requested emodes added
   //   by the library rather than being added here
 
+  ierr = CeedQFunctionSetUserFlopsEstimate(qf, 0); CeedChk(ierr);
+
   // Context data
   CeedQFunctionContext ctx;
   IdentityCtx ctx_data = {.size = 1};