+++ include/ceed-impl.h
@@ -246,6 +246,7 @@ struct CeedQFunction_private {
   CeedQFunctionField *output_fields;
   CeedInt num_input_fields, num_output_fields;
   CeedQFunctionUser function;
+  CeedInt user_flop_estimate;
   const char *source_path;
   const char *kernel_name;
   const char *gallery_name;