+++ doc/sphinx/source/releasenotes.md
@@ -9,6 +9,8 @@ for each release of libCEED.
 
 ### Interface changes
 
+- Added {c:func}`CeedQFunctionSetUserFlopsEstimate` and {c:func}`CeedOperatorGetFlopsEstimate` to facilitate estimating FLOPs in operator application.
+
 (v0-10)=
 
 ## v0.10 (Mar 21, 2022)